🧑‍💻 Refactor script objects' name props & prefab namespace prop
The new changes force imperfect, mistake-prone developers (like me) to be explicit about whether they want the namespace as part of the script name or not.

The type also now provides the namespace name for easy access, instead of having to build it every time you want to—I don't know—use it as the index for an object or some other niche and never-before-seen use case.