🧑‍💻 Introduce SuspenseIfDevelopment component
This component is useful for cases where promises would be fulfilled at buildtime in prod but may not be fulfilled yet in development.

This is particularly useful in cases like function signature tooltips. For all tooltips, we place a Suspense boundary around the tooltip in prod to prevent a 1-frame loading indicator. However, because of this Suspense boundary, if the contents of the tooltip have a pending promise (like a mediawiki data promise), it can seem like the tooltip component is broken.